Move co-sim runtime to Carla/, elevate shared utils to scripts/
The co-sim runtime is a first-class module, not a test artifact: helper_scripts +
run_cosim moved from tests/Sumo/Carla to FIXS_root/Carla/. Reusable utilities
(trafficlight_helper, extract_sumo_tls_as_table, unreal_remove_tl) elevated to
FIXS_root/scripts/. tests/Sumo/Carla now holds only the tests + fixtures, importing
from ../../../Carla. unreal_placing_tls.py imports trafficlight_helper from scripts/.
Dispatch bundles Carla/ + scripts utils preserving the import layout.
