+++ Carla/helper_scripts/run_synchronization/sumo_integration/bridge_helper.py
