+++ Carla/helper_scripts/run_synchronization/sumo_integration/carla_simulation.py
