+++ Carla/helper_scripts/run_synchronization/sumo_integration/sumo_carla_tl_table_manager.py
