+++ Carla/helper_scripts/run_synchronization/sumo_integration/sumo_simulation.py
